requirements.txt
google-api-python-client==2.107.0
oauth2client==4.1.3
Telethon==1.32.0
httplib2~=0.22.0
python-dotenv~=1.0.0
pybit~=5.6.2